archive/v2_langchain/menu_bot_langchain/requirements.txt
# Menu Bot LangChain - Requirements

# Core LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# LangGraph for stateful workflows
langgraph>=0.0.20

# Original dependencies
openai>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pyarrow>=14.0.0

# Sentiment analysis
transformers>=4.30.0
torch>=2.0.0

# Optional: for token analysis
tiktoken>=0.5.0

# Web interface (Streamlit)
streamlit>=1.28.0

# Development tools
pytest>=7.4.0
black>=23.0.0
